+++ src/main/kotlin/raytracer/math/Vector.kt
@@ -2,18 +2,18 @@ package raytracer.math
 
 import kotlin.math.sqrt
 
-data class Vector(val x: Float, val y: Float, val z: Float) {
+data class Vector(val x: Double, val y: Double, val z: Double) {
 
     val normalize: Vector
         get() = this / magnitude
 
-    val magnitude: Float get() = sqrt(x * x + y * y + z * z)
+    val magnitude: Double get() = sqrt(x * x + y * y + z * z)
 
     inline operator fun plus(rhs: Vector) = Vector(x + rhs.x, y + rhs.y, z + rhs.z)
     inline operator fun minus(rhs: Vector): Vector = Vector(x - rhs.x, y - rhs.y, z - rhs.z)
     inline operator fun unaryMinus(): Vector = Vector(-x, -y, -z)
-    operator fun times(scalar: Float): Vector  = Vector(x * scalar, y * scalar, z * scalar)
-    operator fun div(scalar: Float): Vector = Vector(x / scalar, y / scalar, z / scalar)
+    operator fun times(scalar: Double): Vector  = Vector(x * scalar, y * scalar, z * scalar)
+    operator fun div(scalar: Double): Vector = Vector(x / scalar, y / scalar, z / scalar)
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
@@ -23,7 +23,7 @@ data class Vector(val x: Float, val y: Float, val z: Float) {
     }
 }
 
-inline fun dot(a: Vector, b: Vector): Float {
+inline fun dot(a: Vector, b: Vector): Double {
     return a.x * b.x + a.y * b.y + a.z * b.z
 }
 inline fun cross(a: Vector, b: Vector): Vector {
@@ -31,9 +31,9 @@ inline fun cross(a: Vector, b: Vector): Vector {
 }
 
 inline fun reflect(i: Vector, n: Vector): Vector =
-    i - 2.0f * dot(n, i) * n
+    i - 2.0 * dot(n, i) * n
 
-inline operator fun Float.times(v: Vector) =
+inline operator fun Double.times(v: Vector) =
     Vector(this * v.x, this * v.y, this * v.z)
